requirements.txt
keyboard
pywin32
pyinstaller
pillow
